chore: move repo interface into /domain/repo
