+++ app/src/main/java/com/imeanttobe/consensusapp/di/DataStoreModule.kt
@@ -3,7 +3,7 @@ package com.imeanttobe.consensusapp.di
 import android.content.Context
 import androidx.datastore.core.DataStore
 import com.imeanttobe.consensusapp.Id
-import com.imeanttobe.consensusapp.data.repo.IdRepo
+import com.imeanttobe.consensusapp.domain.repo.IdRepo
 import com.imeanttobe.consensusapp.data.repo.IdRepoImpl
 import com.imeanttobe.consensusapp.data.local.serializer.idDataStore
 import dagger.Binds
