+++ app/src/main/java/com/imeanttobe/consensusapp/ui/dev/DevViewModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.imeanttobe.consensusapp.data.repo.IdRepo
+import com.imeanttobe.consensusapp.domain.repo.IdRepo
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
